third_party/blink/web_tests/http/tests/cache/content-type-ignored-during-revalidation-expected.txt

Certain headers (such as 'content-type') ought to be ignored when resources like stylesheets are revalidated via a 304 response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(document.getElementById('test')).backgroundColor is "rgb(0, 255, 0)"
PASS successfullyParsed is true

TEST COMPLETE

This div should have a green background.
